// src/models/Stock.model.ts
import { APISupplierPartDetail } from "./SupplierPart.model";
import { APILocationDetail } from "./Location.model";
import { APICurrency, APIPartstatus } from "./Utils/Utils.model";

export type APIStock = {
	batch: string;
	belongs_to: number;
	build: number;
	consumed_by: number;
	customer: number;
	delete_on_deplete: boolean;
	expiry_date: string;
	is_building: boolean;
	link: string;
	location: number;
	notes: string;
	owner: number;
	packaging: string;
	part: number;
	purchase_order: number;
	purchase_order_reference: string;
	pk: number;
	quantity: number;
	sales_order: number;
	sales_order_reference: string;
	serial: string;
	status: APIPartstatus;
	status_text: string;
	stocktake_date: string;
	supplier_part: number;
	barcode_hash: string;
	updated: string;
	purchase_price: string;
	purchase_price_currency: APICurrency;
	allocated: number;
	expired: boolean;
	installed_items: number;
	stale: boolean;
	tracking_items: number;
	tags: string[];
};

export type APIPartStock = {
	batch: string;
	belongs_to: number;
	build: number;
	consumed_by: number;
	customer: number;
	delete_on_deplete: boolean;
	expiry_date: string;
	is_building: boolean;
	link: string;
	location: number;
	location_detail: APILocationDetail;
	notes: string;
	owner: number;
	packaging: string;
	part: number;
	purchase_order: number;
	purchase_order_reference: string;
	pk: number;
	quantity: number;
	sales_order: number;
	sales_order_reference: string;
	serial: string;
	status: APIPartstatus;
	status_text: string;
	stocktake_date: string;
	supplier_part: number;
	supplier_part_detail: APISupplierPartDetail | null;
	barcode_hash: string;
	updated: string;
	purchase_price: string;
	purchase_price_currency: APICurrency;
	allocated: number;
	expired: boolean;
	installed_items: number;
	stale: boolean;
	tracking_items: number;
	tags: string[];
};

export type APIStockLocation = {
	batch: String;
	belongs_to: number;
	build: number;
	consumed_by: number;
	customer: number;
	delete_on_deplete: boolean;
	expiry_date: string;
	is_building: false;
	link: string;
	location: number;
	location_detail: APILocationDetail;
	notes: string;
	owner: number;
	packaging: string;
	part: number;
	purchase_order: number;
	pk: number;
	quantity: number;
	sales_order: number;
	serial: string;
	status: APIPartstatus;
	status_text: string;
	stocktake_date: string;
	supplier_part: number;
	supplier_part_detail: APISupplierPartDetail;
	barcode_hash: string;
	updated: string;
	purchase_price: string;
	purchase_price_currency: APICurrency;
	allocated: number;
	expired: boolean;
	installed_items: number;
	stale: boolean;
	tracking_items: number;
	tags: string[];
};

export type MovingStock = {
	batch: String;
	delete_on_deplete: boolean;
	location_detail_pathstring: string;
	pk: number;
	quantity: number;
	serial: string;
	status: APIPartstatus;
	status_text: string;
	supplier_part_SKU: string;
	barcode_hash: string;
};

export type PartStock = {
	batch: string;
	build: number;
	expiry_date: string;
	is_building: boolean;
	LocationName: string | null;
	packaging: string;
	part: number;
	pk: number;
	quantity: number;
	serial: string;
	status: APIPartstatus;
	status_text: string;
	supplier_part_detail: {
		name: string;
		url: string | null;
	} | null;
	barcode_hash: string;
	updated: string;
	purchase_price: string;
	purchase_price_currency: APICurrency;
	allocated: number;
	expired: boolean;
	installed_items: number;
	stale: boolean;
	tracking_items: number;
	tags: string[];
};